requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
pyyaml>=6.0.0
playwright>=1.40.0
pyarrow>=14.0.0
tqdm>=4.66.0
lxml>=4.9.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
plotly>=5.15.0
